[cmd/brew-update-forks] push_designated_branches_to_personal_fork_of_active_re- pository:  Change the handling for errors encountered during early filesystem and Git operations.
     The previous error-handling strategy, which incorrectly used `continue`,
was mistakenly copied from within the context of a `for` loop when I was still
using some `brew` internals and/or other Homebrew external commands as examples
to borrow code from.  The new handling uses `return` and an explict argument to
that shell built-in command in each place it's needed.